+++ test/fixtures/sample_many_features_project/lib/features/billing/presentation/billing_page.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/material.dart';
+
+class BillingPage extends StatelessWidget {
+  const BillingPage({super.key});
+  @override
+  Widget build(BuildContext context) => const Scaffold();
+}